Cleans up the unityped languages.
This splits the tests out into different areas
of functionality.

It also changes the `value` function so that it
just uses pattern matches rather than the overkill
that I had before.  The overkill might return, but
not until I get to the point where I'm introducing
modularity into this thing.