Makes the tests for the semantics more precise.
For the small step semantics, we wanted an exclusive or
between a term being a value or having a unique step it could
take rather than an inclusive or.

For the big step semantics, most of the values are part of
the semantics, so we're only after a unique step regardless
of whether the term is a value or not.